// src/proto/cormoran/os_detection/os_detection.ts
export const Os = {
  OS_UNSPECIFIED: 0,
  OS_UNKNOWN: 1,
  OS_WINDOWS: 2,
  OS_MACOS: 3,
  OS_LINUX: 4,
  OS_IOS: 5,
  OS_ANDROID: 6,
  UNRECOGNIZED: -1,
} as const;

export type Os = typeof Os[keyof typeof Os];
// ...
  export type OS_UNSPECIFIED = typeof Os.OS_UNSPECIFIED;
  export type OS_UNKNOWN = typeof Os.OS_UNKNOWN;
  export type OS_WINDOWS = typeof Os.OS_WINDOWS;
  export type OS_MACOS = typeof Os.OS_MACOS;
  export type OS_LINUX = typeof Os.OS_LINUX;
  export type OS_IOS = typeof Os.OS_IOS;
  export type OS_ANDROID = typeof Os.OS_ANDROID;
  export type UNRECOGNIZED = typeof Os.UNRECOGNIZED;